feat: 重构服务层代码并优化数据结构
refactor(post): 将分页参数从i32改为i64以支持更大范围
feat(comment): 添加对评论回复功能的支持
refactor(notification): 移除不必要的UpdatedAt字段返回
feat: 为各服务添加repository层实现
refactor: 统一服务初始化逻辑，添加中间件和日志配置
style: 调整代码格式和import顺序
perf: 实现Redis缓存层以提高性能
fix: 修复like和follow服务的逻辑错误
docs: 更新thrift接口文档
test: 添加repository层单元测试

+++ api-gateway/handler/post.go
@@ -96,8 +96,8 @@ func GetPostList(ctx context.Context, c *app.RequestContext) {
 
 	// 构建请求
 	req := &post.GetPostListRequest{
-		Page:     int32(page),
-		PageSize: int32(pageSize),
+		Page:     page,
+		PageSize: pageSize,
 	}
 
 	if userIDStr != "" {
